requirement.txt
anyio==4.11.0
asgiref==3.10.0
boto3==1.41.5
botocore==1.41.5
certifi==2025.10.5
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.7
django-redis==6.0.0
django-sslserver==0.22
django-storages==1.14.6
djangorestframework==3.16.1
h11==0.16.0
httptools==0.7.1
idna==3.11
inflection==0.5.1
jmespath==1.0.1
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
s3transfer==0.15.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
